fixed BByte1.get() returning float

+++ src/main/java/de/linusdev/lutils/math/vector/abstracts/byten/Byte1.java
@@ -20,7 +20,7 @@ import de.linusdev.lutils.math.vector.abstracts.vectorn.Vector1;
 
 public interface Byte1 extends ByteN, Vector1 {
 
-    default float get() {
+    default byte get() {
         return get(0);
     }
